requirements.txt
Django==4.0.1
django-heroku
gunicorn
django-filter
djangorestframework
drf-nested-routers
celery==4.4.7 
redis